backend/pyproject.toml
[project]
name = "dailyplan-backend"
version = "0.1.0"
description = "FastAPI backend for DailyPlan"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13.2",
  "bcrypt>=4.0.1,<4.1.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.0",
  "openai>=1.30.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.1",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "sqlalchemy>=2.0.30",
  "sqlmodel>=0.0.22",
  "uvicorn[standard]>=0.30.0",
]

[tool.uv]
dev-dependencies = []
